requirements.txt
websocket-client
paho-mqtt